requirements.txt
asgiref==3.4.1
bitrix24-rest==1.1.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Django==4.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
requests==2.26.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
